+++ packages/firebase-hooks/src/firestore/useBatch/batchAdd.js
@@ -1,0 +1,10 @@
+const batchAdd = ({ db, collection, data }) => {
+    const batch = db.batch();
+    data.forEach((item) => {
+        const docRef = db.collection(collection).doc();
+        batch.set(docRef, item);
+    });
+    return batch.commit();
+};
+
+export default batchAdd;