+++ packages/firebase-hooks/src/firestore/useBatch/batchDelete.js
@@ -1,0 +1,10 @@
+const batchDelete = ({ db, collection, data }) => {
+    const batch = db.batch();
+    data.forEach(({ id }) => {
+        const docRef = db.collection(collection).doc(id);
+        batch.delete(docRef);
+    });
+    return batch.commit();
+};
+
+export default batchDelete;