+++ packages/hooks/src/useMutationFlow/index.js
@@ -1,4 +1,3 @@
-import { useCallback } from 'react';
 import { showAppSpinner, hideAppSpinner, showAppNotification } from '@rainbow-modules/app';
 
 const defaults = {
@@ -46,14 +45,14 @@ const useMutationFlow = (opts) => {
         mutation = async () => {},
         successMessage = defaults.success.description,
         errorMessage = defaults.error.description,
-        onSucess = () => {},
+        onSuccess = () => {},
         onError = () => {},
     } = opts;
-    const mutate = useCallback(async (...args) => {
+    const mutate = async (...args) => {
         showAppSpinner();
         try {
             const res = await mutation(...args);
-            onSucess(res);
+            onSuccess(res);
             hideAppSpinner();
             showAppNotification(resolveMessage(successMessage, res, 'success'));
         } catch (error) {
@@ -61,8 +60,7 @@ const useMutationFlow = (opts) => {
             hideAppSpinner();
             showAppNotification(resolveMessage(errorMessage, error, 'error'));
         }
-        // eslint-disable-next-line react-hooks/exhaustive-deps
-    }, []);
+    };
     return { mutate };
 };