+++ packages/forms/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.22.0](https://github.com/nexxtway/rainbow-modules/compare/v0.21.0...v0.22.0) (2021-01-08)
+
+**Note:** Version bump only for package @rainbow-modules/forms
+
+# [0.21.0](https://github.com/nexxtway/rainbow-modules/compare/v0.20.2...v0.21.0) (2021-01-03)
+
+### Bug Fixes
+
+-   fixed multiselect in universal form modal example ([#278](https://github.com/nexxtway/rainbow-modules/issues/278)) ([4e7bb85](https://github.com/nexxtway/rainbow-modules/commit/4e7bb850f1ed88516f97af7865b85891973599ab))
+
+## [0.20.2](https://github.com/nexxtway/rainbow-modules/compare/v0.20.1...v0.20.2) (2021-01-01)
+
+### Bug Fixes
+
+-   useMutationFlow type in onSuccess prop ([9aeabe0](https://github.com/nexxtway/rainbow-modules/commit/9aeabe0b54ec3ac2e07ebc1cec04e507dfb8ed4a))
+
+## [0.20.1](https://github.com/nexxtway/rainbow-modules/compare/v0.20.0...v0.20.1) (2020-12-31)
+
+**Note:** Version bump only for package @rainbow-modules/forms
+
+# [0.20.0](https://github.com/nexxtway/rainbow-modules/compare/v0.19.2...v0.20.0) (2020-12-30)
+
+### Bug Fixes
+
+-   fixed universal form modal example ([#271](https://github.com/nexxtway/rainbow-modules/issues/271)) ([4839327](https://github.com/nexxtway/rainbow-modules/commit/483932732669d6457a9c21cb71dafbae50bdf6fd))
+-   universal form modal examples ([#273](https://github.com/nexxtway/rainbow-modules/issues/273)) ([eda9ff0](https://github.com/nexxtway/rainbow-modules/commit/eda9ff0d221f9c83f3db3d9f810e33626c2a2ef6))
+
 ## [0.19.2](https://github.com/nexxtway/rainbow-modules/compare/v0.19.1...v0.19.2) (2020-12-28)
 
 ### Bug Fixes